+++ nullstone.tf
@@ -3,6 +3,12 @@ terraform {
     ns = {
       source = "nullstone-io/ns"
     }
+    google = {
+      source  = "hashicorp/google"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
+    }
   }
 }